refactor: 응원하기 버그 값 수정
- int -> long 타입으로 변경
- 1,000이 넘는 값이 들어왔을 때, 1,000으로 치환해서 저장하도록 수정

+++ backend/src/main/java/feedzupzup/backend/organization/domain/Organization.java
@@ -65,7 +65,7 @@ public class Organization extends BaseTimeEntity {
         this.cheeringCount.add(other);
     }
 
-    public int getCheeringCountValue() {
+    public long getCheeringCountValue() {
         return cheeringCount.getValue();
     }
