+++ backend/src/main/java/feedzupzup/backend/organization/dto/request/CheeringRequest.java
@@ -6,7 +6,7 @@ import io.swagger.v3.oas.annotations.media.Schema;
 @Schema(description = "응원하기 요청")
 public record CheeringRequest(
         @Schema(description = "응원한 개수", example = "200")
-        int cheeringCount
+        long cheeringCount
 ) {
 
         public CheeringCount toCheeringCount() {
